Somewhere im looping too much

+++ d2q9-bgk.c
@@ -436,7 +436,7 @@ float av_velocity(const t_param params, t_speed* cells, int* obstacles, t_ocl oc
       tot_cells += sum_cells[x];
       tot_u += sum_u[x];
   }
-  printf("Num_wkg: %d\n", params.num_wkg);
+  printf("Num_wkg: %d\n", (int)params.num_wkg);
   free(sum_cells);
   free(sum_u);
   // printf("%d\n", tot_cells);
